Camera Displacement
Created a camera displacement variable that is used in the solution base class to update the camera on the focuspoint. This allows various controls to update a single value that is used to update camera location in the animation loop.

Clean up needed with solution classes logic to ensure proper encapsulation.